# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tark API
  version: 1.0.0
  description: |
    The contract between the Tark app and the Tark backend (Go, PostgreSQL,
    hosted on ArvanCloud). Both sides are written against this file; neither
    one is the source of truth on its own.

    It covers accounts and sign-in (email + password with a verification
    code or link, Google Sign-In with account linking, password reset and
    change, sessions, and the email-change flow the app does not offer yet),
    the profile, and the subscription surface the app's offline-first
    entitlement logic depends on.

    Conventions
    - Every endpoint is HTTPS only and versioned under /v1.
    - Errors use one shape (Problem). `code` is stable and machine-readable;
      the app never shows `detail` to people.
    - Writes that the app may retry carry an `Idempotency-Key` header. The
      same key with the same body returns the first response again; the same
      key with a different body is refused with 422 `idempotency_mismatch`.
    - Times are Unix epoch milliseconds (int64), always from the server clock.
servers:
  - url: https://api.tarkk.ir/v1
security:
  - accessToken: []
paths:
...
  /auth/register/verify:
    post:
      operationId: verifyRegistration
      summary: Verify the code or link; creates the account and signs in.
      description: |
        The app calls this as soon as the last digit is typed, or as soon
        as the link opens the app. Success signs the app in directly.

        Failures: `code_invalid` (422, with `attemptsLeft`), `code_locked`
        (423, after 5 wrong codes; resend for a new one), `flow_expired`
        (410), `flow_not_found` (404), `email_already_registered` (409, the
        address got an account meanwhile; sign in instead).

        Repeating a successful verify within 10 minutes (the answer got
        lost) signs in again instead of failing.
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/OptionalInstallKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofRequest'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '423':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
  /auth/google/nonce:
    post:
      operationId: googleNonce
      summary: A one-time nonce to pass to the Google SDK before sign-in.
      description: |
        Google copies it into the ID token. The server accepts each nonce
        once and only for 10 minutes, so an ID token captured elsewhere or
        replayed is refused.
      security: []
      responses:
        '200':
          description: The nonce.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleNonce'
        '429':
          $ref: '#/components/responses/Problem'
  /auth/google:
    post:
      operationId: googleSignIn
      summary: Sign in or sign up with a Google ID token.
      description: |
        The server verifies the ID token itself (signature against Google's
        keys, issuer, audience, expiry, verified email, nonce) and takes the
        email and Google account id only from it.

        - Known Google account: 200, signed in.
        - The address already has a password account: 409 `link_required`
          with `ticket`, `ticketExpiresAt` and a masked `email`. Ask for that
          account's password and call /auth/google/link.
        - New address: account created with `name` from the request (the
          name the app already has locally) or, if absent, Google's name.
          If neither exists, 422 `name_required` with `ticket` and maybe
          `suggestedName`; ask for a name and call /auth/google/complete.

        Other failures: 401 `google_token_invalid` (get a fresh token),
        422 `google_email_unverified`, 409 `account_conflict` (the address
        is on an account linked to a different Google account),
        409 `retry_sign_in` (a race; sign in with Google again),
        503 `google_unavailable` (Google's keys could not be fetched).
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/OptionalInstallKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignInRequest'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          description: '`link_required`, `account_conflict` or `retry_sign_in`.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/GoogleTicketProblem'
        '422':
          description: '`name_required` or `google_email_unverified`.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/GoogleTicketProblem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /auth/google/link:
    post:
      operationId: linkGoogle
      summary: Link Google to an existing password account, after its password.
      description: |
        Wrong password: 401 `invalid_credentials` with `attemptsLeft`. The
        ticket allows 5 tries (then 423 `ticket_locked`) and shares the login
        failure limits. 410 `ticket_expired`: sign in with Google again.
        Success signs in and emails the account that Google was added.
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/OptionalInstallKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLinkRequest'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '423':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /auth/google/complete:
    post:
      operationId: completeGoogleSignup
      summary: Finish a Google sign-up that needed a name.
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/OptionalInstallKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleCompleteRequest'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
  /auth/logout-all:
    post:
      operationId: logoutAll
      summary: End every session of the account, this one included.
      responses:
        '204':
          description: Signed out everywhere.
        '401':
          $ref: '#/components/responses/Problem'
  /auth/password/forgot:
    post:
      operationId: forgotPassword
      summary: Start a password reset. Same answer for every address.
      description: Only an address with an account receives a code; others receive nothing.
      security: []
      parameters:
        - $ref: '#/components/parameters/OptionalIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotRequest'
      responses:
        '202':
          $ref: '#/components/responses/FlowStarted'
        '400':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
  /auth/password/forgot/verify:
    post:
      operationId: verifyReset
      summary: Verify the reset code or link; returns a ticket for the new password.
      description: Same failures as /auth/register/verify. Verifying again replaces the ticket.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofRequest'
      responses:
        '200':
          description: The ticket for /auth/password/reset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetTicket'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '423':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /auth/password/reset:
    post:
      operationId: resetPassword
      summary: Set the new password and sign in. Every other session ends.
      description: |
        A Google-only account gains a password this way. The account's
        email is told the password changed. 410 `ticket_expired` when the
        ticket is used, old or unknown. 422 with a `password_*` code when
        the password is refused.
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/OptionalInstallKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetRequest'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /auth/password/change:
    post:
      operationId: changePassword
      summary: Change the password, given the current one.
      description: |
        This session stays signed in; every other session ends. 401
        `invalid_credentials` for a wrong current password. 409
        `password_not_set` for a Google-only account (use forgot password
        to add one). 422 with a `password_*` code or `password_unchanged`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Changed.
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /auth/email-change:
    post:
      operationId: startEmailChange
      summary: Start changing the account's email (not offered by the app yet).
      description: |
        Needs fresh proof: the current password, or for a Google-only
        account a fresh Google ID token of the linked Google account. A code
        and link go to the new address; nothing changes until it is
        verified.
      x-app-status: not used by the app yet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailChangeRequest'
      responses:
        '202':
          $ref: '#/components/responses/FlowStarted'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
  /auth/email-change/verify:
    post:
      operationId: confirmEmailChange
      summary: Verify the new address; it becomes the account's email.
      description: |
        The old address is told; every other session ends. 409
        `email_in_use` when the address belongs to another account.
      x-app-status: not used by the app yet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofRequest'
      responses:
        '200':
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '423':
          $ref: '#/components/responses/Problem'

  # ---- Profile -------------------------------------------------------------
  /profile:
    get:
      operationId: getProfile
      summary: The signed-in account's profile.
      responses:
        '200':
          description: The profile. The ETag header is its version, for If-Match on PUT.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Problem'
    put:
      operationId: putProfile
      summary: Replace the editable profile fields.
      description: |
        Editable: name and avatarId. Email is read-only here; any other
        field in the body is ignored. With If-Match, a write based on a
        stale copy (another device changed the profile) is refused with
        412 `profile_changed` instead of silently undoing that change.
        Sending the current values again always succeeds, so retries are
        safe.
      parameters:
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: The updated profile.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /account/delete:
    post:
      operationId: deleteAccount
      summary: Delete the account and its data, at once and for good.
      description: |
        A "yes" is not enough. The request needs all of:
        - `confirmEmail`: the account's email, typed by the person (case
          and surrounding spaces do not matter). Anything else is 422
          `confirmation_mismatch`.
        - fresh proof: `currentPassword`, or for a Google-only account a
          new `googleIdToken` of the linked Google account (get a nonce
          first). Wrong proof is 401 `invalid_credentials`.
        - while a paid Bazaar period is running, `subscriptionAcknowledged:
          true`. Without it the answer is 409 `subscription_active` with
          `autoRenewing`, so the app can say that deleting the account does
          not cancel the Bazaar subscription.

        On success every session ends, a confirmation email goes to the
        address, and the email can be used for a new account. Bazaar
        purchases are removed with the account, so the same purchase can
        be restored in a new account later.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '204':
          description: Deleted. The tokens this app holds no longer work.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'

  # ---- Subscription --------------------------------------------------------
  /subscription:
    get:
      operationId: getSubscription
      summary: Current subscription state as a signed entitlement.
      description: |
        Returns the caller's subscription as the server knows it right now,
        re-checking Cafe Bazaar first when the stored state is stale.

        A Bazaar outage or timeout never downgrades anyone: the server answers
        with the last state it verified and sets `bazaarChecked: false`.
      parameters:
        - $ref: '#/components/parameters/InstallKey'
      responses:
        '200':
          description: The caller's subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      description: Short-lived access token from the sign-in endpoints.
  parameters:
    Platform:
      name: X-Tark-Platform
      in: header
      required: false
      description: The platform signing in. Sign-in exists on Android and iOS only.
      schema:
        type: string
        enum: [android, ios]
    OptionalInstallKey:
      name: X-Tark-Install-Key
      in: header
      required: false
      description: The install's public key (as for /subscription), stored on the new session.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{43}$'
...
  responses:
...
    Session:
      description: Signed in. Carries the tokens and the profile, so the app can show the signed-in state at once.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Session'
    Problem:
      description: A failure, in the one shape every endpoint uses.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    ProofRequest:
      type: object
      required: [flowId]
      description: Exactly one of code or linkToken.
      properties:
        flowId:
          type: string
        code:
          type: string
          pattern: '^[0-9]{6}$'
        linkToken:
          type: string
          description: The part after `#` in the email link.
...
    GoogleNonce:
      type: object
      required: [nonce, expiresAt]
      properties:
        nonce:
          type: string
        expiresAt:
          type: integer
          format: int64
    GoogleSignInRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          maxLength: 8192
        name:
          $ref: '#/components/schemas/Name'
        locale:
          $ref: '#/components/schemas/Locale'
...